% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_matches}
\alias{all_matches}
\title{Point level dataset - all matches}
\format{
A dataframe:

\itemize{
\item set: the current set of the match (1-5 for ATP matches and 1-3 for WTA matches)
\item player1_game_score: score of the current game for player1
\item player2_game_score: score of the current game for player2
}
}
\usage{
all_matches
}
\description{
This is a dataframe with all matches from RG 2019 - 2024
}
\keyword{datasets}
