% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_combine.R
\name{clean_and_combine}
\alias{clean_and_combine}
\title{Clean and Combine Function}
\usage{
clean_and_combine(
  list_of_matches,
  player_interest = "(.|\\\\s)*\\\\S(.|\\\\s)*"
)
}
\arguments{
\item{list_of_matches}{is a list of all matches of interest - from the fetch_all_matches() function}

\item{player_interest}{is a string of the player's name we want as player1 - first or last name (case insensitive)}
}
\value{
returns a cleaned data frame with all matches of interest (point level of granularity)
}
\description{
This is a function that cleans the Court Vision data and combines all matches of interest into a data frame.
The clean_point_level() function is used to clean one match at a time.
}
\examples{
nadal_2022 <- fetch_all_matches(player = "Nadal", year = 2022)
clean_and_combine(nadal_2022, player_interest = "Nadal")

}
