% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_shot_level.R
\name{clean_shot_level}
\alias{clean_shot_level}
\title{Clean Shot Level}
\usage{
clean_shot_level(raw_data)
}
\arguments{
\item{raw_data}{is a data frame of the raw Court Vision data (must be output from the fetch_all_matches function)}
}
\value{
a data frame with several rows (hit, net, peak, bounce) for each shot in the match/matches of interest
}
\description{
This is a function that cleans the Court Vision data to the \emph{shot} level of granularity.
}
\examples{
nadal_final_2022 <- fetch_all_matches(player = "Nadal", year = "2022", round = "F")
clean_shot_level(nadal_final_2022)

}
