% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_court.R
\name{draw_court}
\alias{draw_court}
\title{Draw Court}
\usage{
draw_court()
}
\value{
returns a ggplot object with solid lines representing the lines on the tennis court, dashed line represents the net
}
\description{
This is a function that draws the tennis court (dimensions are to scale)
}
\examples{
draw_court()

}
