% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all_matches.R
\name{fetch_all_matches}
\alias{fetch_all_matches}
\title{Fetch All Matches}
\usage{
fetch_all_matches(
  player = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  year = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  round = "(.|\\\\s)*\\\\S(.|\\\\s)*"
)
}
\arguments{
\item{player}{is a string of the player's name - first or last name (case insensitive)}

\item{year}{is a string of the year the match was played - between 2019 and 2023}

\item{round}{is a string of the round the match was played - F, SF, QF, R16, R32, R64, or R128}
}
\value{
a data frame of all matches that match the player, year, and/or round of interest
}
\description{
This is a function that collects and binds all matches that match the player, year, and/or round of interest.
}
\examples{
fetch_all_matches(player = "Nadal", year = "2022", round = "F") ## fetches the 2022 French Open final match that Nadal played
fetch_all_matches(player = "Federer") ## fetches all matches that Federer played

}
