% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_match_info.R
\name{fetch_match_info}
\alias{fetch_match_info}
\title{Fetch Match Info}
\usage{
fetch_match_info(
  year = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  player_last = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  round = "(.|\\\\s)*\\\\S(.|\\\\s)*"
)
}
\arguments{
\item{year}{is a string of the year the match was played - between 2019 and 2023}

\item{round}{is a string of the round the match was played - F, SF, QF, R16, R32, R64, or R128}

\item{player}{is a string of the player's name - first or last name (case insensitive)}
}
\value{
a list of strings with the info of all matches that match the player, year, and/or round of interest
}
\description{
This is a function that fetches all matches that match the player, year, and/or round of interest.
}
\examples{
fetch_match_info(player = "Nadal", year = "2022", round = "F") ## fetches the 2022 French Open final match that Nadal played
fetch_match_info(player = "Federer") ## fetches all matches that Federer played

}
