% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matches.R
\name{filter_matches}
\alias{filter_matches}
\title{Filter Matches Function}
\usage{
filter_matches(
  player = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  year = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  round = "(.|\\\\s)*\\\\S(.|\\\\s)*"
)
}
\arguments{
\item{player}{is a string of the player's name - first and last name (case sensitive)}

\item{year}{is a string of the year the match was played - between 2019 and 2023}

\item{round}{is a string of the round the match was played - F, SF, QF, R16, R32, R64, or R128}
}
\value{
returns a point-level data frame of all matches given the specified player, year, and/pr round
}
\description{
This is a function that finds all matches of a specified player, year, and/or round
}
