% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_shot_level.R
\name{clean_shot_level}
\alias{clean_shot_level}
\title{Clean Shot Level Function}
\usage{
clean_shot_level(cleaned_data)
}
\arguments{
\item{cleaned_data}{is a data frame of cleaned point-level data}
}
\value{
returns a data frame with several rows (hit, net, peak, bounce) for each shot in the match/matches of interest
}
\description{
This is a function that parses the trajectory data from the Court Vision data - breaking the match/matches down to the \emph{shot} level of granularity.
}
\examples{
nadal_2022_cleaned <- clean_and_combine(nadal_2022, player_interest = "Nadal")
clean_shot_level(nadal_2022_cleaned)

}
