% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matches.R
\name{filter_matches}
\alias{filter_matches}
\title{Filter Matches Function}
\usage{
filter_matches(
  player = "(.|\\\\s)*\\\\S(.|\\\\s)*",
  year_of_interest = "(.|\\\\s)*\\\\S(.|\\\\s)*"
)
}
\arguments{
\item{player}{is a string of the player's name - first and last name (case sensitive)}

\item{year_of_interest}{is a string of the year the match was played - between 2019 and 2023}
}
\value{
returns a point-level data frame of all matches given the specified player and year
}
\description{
This is a function that finds all matches of a specified player, year, and/or round
}
